requirements.txt
Flask==2.2.2
openai==0.25.0
python-dotenv==0.21.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1